requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
cryptography==2.3.1
Fabric3==1.14.post1
idna==2.7
invoke==1.2.0
paramiko==2.4.2
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
six==1.11.0
